Move disk caching to its own class
